+++ test/tc_helper.rb
@@ -9,9 +9,11 @@ SimpleCov.start do
 end
 
 require 'minitest/autorun'
+require 'rspec/mocks/minitest_integration'
 require 'timecop'
 require 'webmock/minitest'
 require 'axlsx'
+require 'ooxml_crypt' if RUBY_ENGINE == 'ruby'
 
 module Minitest
   class Test
@@ -45,6 +47,10 @@ module Minitest
 
     def truffleruby?
       RUBY_ENGINE == 'truffleruby'
+    end
+
+    def ooxml_crypt_available?
+      defined?(OoxmlCrypt)
     end
   end
 end